wip: integrate oid4vp flow into wallet
