+++ Cargo.toml
@@ -15,11 +15,15 @@ name = "mobile_sdk_rs"
 name = "uniffi-bindgen"
 path = "uniffi-bindgen.rs"
 
+[features]
+default = ["reqwest"]
+reqwest = ["oid4vp/reqwest"]
+
 [dependencies]
 uniffi = { version = "0.27.1", features = ["cli"] }
 
 isomdl = { git = "https://github.com/spruceid/isomdl", rev = "1f4f762" }
-
+oid4vp = { path = "../oid4vp-rs", default-features = false }
 thiserror = "1.0.56"
 hex = "0.4.3"
 p256 = { version = "0.13.2", features = ["pkcs8"] }
@@ -28,6 +32,8 @@ serde_cbor = "0.11.2"
 serde_json = "1.0.111"
 uuid = { version = "1.6.1", features = ["v4"] }
 tracing = "0.1.40"
+async-trait = "0.1.81"
+url = "2.5.2"
 
 [dev-dependencies]
 uniffi = { version = "0.27.1", features = ["bindgen-tests"] }
