+++ src/common.rs
@@ -1,0 +1,27 @@
+// Re-export Uuid for common use.
+pub use url::Url;
+pub use uuid::Uuid;
+
+use crate::UniffiCustomTypeConverter;
+
+uniffi::custom_type!(Uuid, String);
+impl UniffiCustomTypeConverter for Uuid {
+    type Builtin = String;
+    fn into_custom(uuid: Self::Builtin) -> uniffi::Result<Self> {
+        Ok(uuid.parse()?)
+    }
+    fn from_custom(uuid: Self) -> Self::Builtin {
+        uuid.to_string()
+    }
+}
+
+uniffi::custom_type!(Url, String);
+impl UniffiCustomTypeConverter for Url {
+    type Builtin = String;
+    fn into_custom(url: Self::Builtin) -> uniffi::Result<Self> {
+        Ok(Url::parse(&url)?)
+    }
+    fn from_custom(url: Self) -> Self::Builtin {
+        url.to_string()
+    }
+}
